fix: 코드리뷰 관련 수정
